task/-/Implementing issue showing type change to svelte [Task] - issue 보여주는 타입을 svelte로 구현
